+++ tests/testsqlite.nim
@@ -14,8 +14,10 @@ db(dbName, "", "", ""):
   type
     User {.table: "users".} = object
       email {.unique.}: string
+      ssn: Option[int]
       birthDate {.
         dbType: "INTEGER",
+        notNull,
         parseIt: it.i.fromUnix().local(),
         formatIt: dbValue(it.toTime().toUnix())
       .}: DateTime
@@ -34,6 +36,7 @@ db(dbName, "", "", ""):
       book {.
         dbCol: "bookId",
         dbType: "INTEGER",
+        notNull,
         fk: Book
         parser: getBookById,
         formatIt: dbValue(it.id),
@@ -47,7 +50,7 @@ suite "Creating and dropping tables, CRUD":
 
       for i in 1..9:
         var
-          user = User(email: "test-$#@example.com" % $i,
+          user = User(email: "test-$#@example.com" % $i, ssn: some i,
                       birthDate: parse("200$1-0$1-0$1" % $i, "yyyy-MM-dd"))
           publisher = Publisher(title: "Publisher $#" % $i)
           book = Book(title: "Book $#" % $i, authorEmail: user.email,
@@ -70,20 +73,21 @@ suite "Creating and dropping tables, CRUD":
       let query = "PRAGMA table_info($#);"
 
       check dbConn.getAllRows(sql query % "users") == @[
-        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 1],
-        @[dbValue 1, dbValue "email", dbValue "TEXT", dbValue 0, dbValue nil, dbValue 0],
-        @[dbValue 2, dbValue "birthDate", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 0]
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "email", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "ssn", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 0],
+        @[dbValue 3, dbValue "birthDate", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 0]
       ]
       check dbConn.getAllRows(sql query % "books") == @[
-        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 1],
-        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 0, dbValue nil, dbValue 0],
-        @[dbValue 2, dbValue "authorEmail", dbValue "TEXT", dbValue 0, dbValue nil, dbValue 0],
-        @[dbValue 3, dbValue "publisherTitle", dbValue "TEXT", dbValue 0, dbValue nil, dbValue 0],
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "authorEmail", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 3, dbValue "publisherTitle", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
       ]
       check dbConn.getAllRows(sql query % "editions") == @[
-        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 1],
-        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 0, dbValue nil, dbValue 0],
-        @[dbValue 2, dbValue "bookId", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 0]
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "bookId", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 0]
       ]
 
   test "Create records":
@@ -227,20 +231,21 @@ suite "Creating and dropping tables, CRUD":
       let query = "PRAGMA table_info($#);"
 
       check dbConn.getAllRows(sql query % "users") == @[
-        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 1],
-        @[dbValue 1, dbValue "email", dbValue "TEXT", dbValue 0, dbValue nil, dbValue 0],
-        @[dbValue 2, dbValue "birthDate", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 0]
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "email", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "ssn", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 0],
+        @[dbValue 3, dbValue "birthDate", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 0]
       ]
       check dbConn.getAllRows(sql query % "books") == @[
-        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 1],
-        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 0, dbValue nil, dbValue 0],
-        @[dbValue 2, dbValue "authorEmail", dbValue "TEXT", dbValue 0, dbValue nil, dbValue 0],
-        @[dbValue 3, dbValue "publisherTitle", dbValue "TEXT", dbValue 0, dbValue nil, dbValue 0],
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "authorEmail", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 3, dbValue "publisherTitle", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
       ]
       check dbConn.getAllRows(sql query % "editions") == @[
-        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 1],
-        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 0, dbValue nil, dbValue 0],
-        @[dbValue 2, dbValue "bookId", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 0]
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "bookId", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 0]
       ]
 
     withCustomDb(customDbName, "", "", ""):